Update to infection severity parameters/assumptions
Incorporation of age disaggregation of infection severity and updated care pathway rates. Also inclusion of death compartment in model.

simtree works intermittently, but sometimes gets stuck at simgeneration() function and so simforest() generally not working. Possibly an issue with the timetransmission values input to the if statement 'if h["timetransmission"] < maxtime'