Update core.jl
Additions to enable the maximum number of infections allowed in the simulation to be allocated pro rata to simtree outputs relating to different importation events. There are two methods for the allocation of the total max_cases:
(1) linear - based on time between importation and the end of the sim (maxtime = maximum time of infection)
(2) exponential - based on an estimate of total cases relating to each importation by the end of the sim using an exponential growth model.
The linear model seems to work better with maxcases = 50000, i.e fewer simtree runs hit their individual limit, and when they do the last infection time (tinf) is closer to maxtime (=90 days)