Create icu_sample_prob_functions.jl
Functions used in the sampling of ICU cases:
- sample_nhs_trust
Samples an NHS Trust code based on the ITL2 region and the age of the infectee.                         The function distinguishes between adult and paediatric ICU admissions, using                        different probability distributions depending on age.
- icu_sample_prob
Calculates the probability of sampling an ICU patient for metagenomic testing based on admission to particular NHS Trust, patient age, pathogen type, and site engagement stage.
- icu_sample_prob_region
Returns probability of a positive metagenomic sample based on home region, age of infectee, pathogen type and stage categorisation of sites to include for testing
- sample_icu_cases
Samples cases based on a proportion of cases, either nationally or in specific regions. The number of background ARI admissions is therefore not explicitly taken into account, but is implicit.
- sample_icu_cases_n
Samples cases based on a number of samples rather than just a proportion of cases. The number of samples is converted to a probability using an estimate of the background number of ICU ARI cases and the modelled pathogen X ICU ARI cases.